add SecretLinkQRCode class
